✨(backend) utility methods for batch order
We will need to trigger manually some methods on the batch
order from the backoffice. In the meantime, those methods will
be used in the django admin to handle a batch order. Those methods
regroup assignation of an organization, submitting to signature the
contract, validating a successful payment, and send the email
with vouchers.