✨(backend) add order `to_own` state in binding order states
We want our orders that are already paid that were generated
by a batch order to be accounted as seats taken on an order
group if one has been setted on the batch order at the
creation. In fact, before adding the `to_own` state into the
binding states, when a batch order would generate its orders,
the number of binding orders on the order group would not
include them into the count.