✨(backend) add order state `to_own` to available seats property
We want our orders that are already paid that were generated
by a batch order to be accounted as seats taken on an order
group if one has been setted on the batch order at their
creation. The property available_seats on the order group
did not count the orders in state `to_own`, we have changed
this behavior to take them in account with the binding states
orders. This allows us to keep track of orders that are related
to the order group, thus knowing exactly how many seats left
are remaining.