✨(backend) admin api batch order generate orders
Once the batch order is in state `completed`, we allow the admin
user to generate the orders if they were not yet generated before.
If the state of the batch order is not in completed state, we don't
allow to generate the orders. If the orders were already generated,
we also block the admin user of this action.
When all the orders are generated, we send the voucher codes by
email to the batch order owner. The generation of orders and
the email are done with a celery task.