✅(backend) fix flaky test order group number of seats
The OrderGroup factory for testing purposes would set
automatically a random number between 0 and 100 for
the number of seats available. When the OrderGroup
factory sets the number of seats to 0, it blocks the
creation of a new order related to this order group.
To avoid this behavior, we have decided change the
random value setted by the factory between 1 and 100.