✨(backend) batch order quote models to prepare document
Some of our users will want to receive a quote before proceeding
to the confirmation of their batch order. In fact, some companies
work with quotes and they don't pay with the credit card or
with a simple SEPA bank transfer. They usually use another platform
and they needs to have the quote of the batch order to generate
it on their side. For this needs, we have a create a model that
allows us to generate on demande quote for a batch order where
it regroups all the essential informations, with estimates prices,
and the content of the product into the active language.