✨(backend) added state `quoted` for batch order flow
When the batch order is related to a quote, it means
that the payment will occur outside our payment backend.
It's a new method of payment for companies. This action
will demand a manual validation from our admin users
once the organization has confirmed that it has published
the purchase order from the quote received.
Once this step is validated, the batch order can transition
to completed.