👔(backend) add properties on batch order model
To help developers while working with batch order objects,
we have added new properties on the model allowing
us to verify quickly if the object is eligible to be or is in a
specific state. It avoids us comparing the object to different
states. Also, it reflects the business logic of the batch order.