✨(backend) payment schedule activate for credential product
We needed to change the payment schedule limits and percentages
because there will be offers that will be lower than the
lowest limit previously set of 150. When it comes to credential
products, we need to always have the first installment to be
30% of the total price. Our previous setting was blocking the
sellers to make offers at less than 150 euros for a course.
Previously, when the price was less than 150 euros, there
would only be 1 installment. We've changed this behaviour
for credential products. Also, we needed to make some
changes for the certificate products because they should
remain a one-off payment, meaning that no matter the price
setted, it should always be 1 installment at the most
representing 100% of the price.

Fix #1115