✨(backend) quote model for batch order payment
We added the Quote model to allow companies to pay
different way a batch order. The Quote is in fact a sort
of small contract that needs to be accepted first
by the organization and then by the buyer. Once
both parties accepts the quote, the buyer can send
a purchase order to the organization. Once this step
is done, we can consider that the batch order will be
paid by the buyer, it's an engagement.