✨(backend) quote and quote definition for batch order
We added the `Quote` model to allow companies to pay
a different way a batch order. The `Quote` is a sort
of small contract that needs to be accepted first
by the organization and then by the buyer. Once
both parties accepts the quote, the buyer can send
a purchase order to the organization. Once this step
is done, we can consider that the batch order will be
paid by the buyer, it's an engagement. Also, we have
created a `QuoteDefinition` that describes the
template and the markdown content when generating
the context for the document.