✨(backend) add discounted price on `ProductRelationSerializer`
Our API client would not inform to our consumers that there
is a discounted price for the purchase of a certificate
when they open the sales tunnel. Actually, even if an offer
rule was activated for the product certificate, it always
mentions the initial product's price and not the discounted
price. The `ProductRelationSerializer` that is used for our
API consumers now sees the discounted price when there is one.
Otherwise, they can rely on the product's initial price.