🐛(backend) generate contract definition pdf aware of timezone
We have discovered while generating the PDF on the server side,
it would ignore the timezone of the user. This caused the
course start and course end values to differ from those set
via the client interface, which uses timezone awareness
datetimes. We now have fixed this issue to not ignore the
configured timezone while generating the PDF contract definition.