🔥(backend) remove `OfferingRule` model
We get rid of the `OfferingRule` model because
we have forgotten to make the field description
translatable. It seems complicated to just add
the field and make the model inherit from parler
models. A lot of conflicts while attempting
to migrate. Since nobody is using this model yet,
we can afford to get rid of it.