✨(backend) generate orders for batch order with quote
When a batch order is related to a quote, only if the quote
has received the purchase order, then we can allow to generate
the orders. Otherwise, if the batch order is related to a quote but
it has not yet received the purchase order, it cannot be possible
to generate the orders yet.