✨(backend) create batch order with specific organization
Before the client API would search for the organization who has
the least orders when it came to create a batch order. Now, the
authenticated user can decide which organization he wants for
his batch order. If the organization is not passed, then we let
the auto-assignation logic in place choose for us the organization
who has the least orders.