✨(backend) preview payment schedule and reduced price with voucher
We want anonymous and authenticated user to be able to
preview the estimated total price of an offering when there
is a voucher code on the purchase of an offering (credential
or certificate). Actually, only the discounted price of the
initial price is returned on the `payment_schedule` endpoint.
We updated this endpoint to return the estimated price with
a voucher code and the payment schedule accordingly.