✨(backend) add payment methods on batch order
Since we have different payment methods for a batch
order, we needed to add this field into the model.
Also, we have changed the behavior in the creation
of a batch order. Now, when we create one, we will
 also relate it to a quote. The total should stay at 0
while waiting for the organization to input the total
value of the batch order (since the price will depend on various
criterias). Only once the organization sets a total, we then
create the main invoice related to the batch order.