✨(backend) confirm organization signature on a quote
For the batch order's quote, an organization needs to
to confirm that they have signed the batch order's quote.
Through this new endpoint, we get the total price for the
batch order since it's B2B offer. With this value, we can
finally set the total on the batch order's price by freezing
the total, and it creates the main invoice.