✨(backend) endpoint to confirm organization signature
For the batch order's quote, an organization needs an endpoint
to confirm that they have signed the batch order and at that time
we receive the total amount set for the batch order. With this value,
we can set the total on the batch order price, and also create the main
invoice.