✨(backend) improve payment schedule endpoint
We need to return into the response of this endpoint
the difference made with the voucher code or
the discounted price on the offering from its initial
price. Now, this endpoint returns respectively, the
initial price, the discount applied if one is set and finally
the discounted price. If no discount, nor voucher code
is set, it return a value for the initial price and sets
None for the other two keys (`discount` and `discounted_amount`).