♻️(backend) refactor batch order to sign flow state
Since we have different payment methods and the signature should
be available only after certain steps, we updated the flow of batch order
and the properties on the batch order model. A batch order with a
payment method of purchase order can only sign the contract once the
purchase order has been received. Whereas, with the two other payment
methods, only once the quote is signed by the organization, then the
submission of the contract to sign is available. We needed to handle those
cases.