✨(backend) confirm bank transfer for batch order
Organizations can confirm that they have received the bank
transfer for a batch order. This endpoint will generate the
creation of the child invoices. Thus, the state will transition
from signing to completed directly because it has been
considered as paid. We have added a method that catches
into the post transition success in the batch order flow which
triggers the creation of the orders and the vouchers. Once
they are generated, they are finally sent by email to the
batch order owner.