✨(backend) batch order flow from signing to completed
Those changements concerns batch order where the payment
method is set with purchase order. When the batch order's
contract is signed by the buyer, the batch order can be
considered as completed because the purchase order
confirm the engagement of the buyer to pay up afterwards
the organization. It's the only process where paying
is before the signature of the contract. By changing the
state to completed into the signature backend, it allows
to trigger the method into the batch order flow
transition success from signing to completed.