✨(backend) micro credential template without unicamp brand
Currently, our only template for the micro credential (also
known as micro certificate) was ready for Unicamp brand.
We want to be able to generate those documents without
the brand. There is now 2 available version of this document,
the default one and the branded one. This development
involved in creating a migration file because we have renamed
the enums of the certificate template names. We also made a
symbolic link for the stylesheet of the
microcredential_degree_unicamp, because for now, they share
the same one as the default one.

Fix #1073