✨(backend) batch order new fields for creation
We need additional information about the buyer for a batch order
concerning the billing address if it's different than the company's
address, and also, we need to get the informations about the
representative of the buyer. Because, in some cases, it might
not be the same one who initiates the purchase.