✨(backend) batch order new fields for creation
We need additional information about the company
who purchases the batch order. Those information
concerns the billing address, the administrative
representative of the company (who will follow up
the signature of the contract). We needed those
informations because it might not be the logged in
user who can have this role, also, the billing address
can differ from the company's address.