✨(backend) quote template and document context preparation
There are new fields about the buyer's company that needs
to be added to the quote. The billing address, funding
entity and the tax registration number were added to
the template. Since we are updating the quote, we have
decided to take away sections that are not used
anymore for the document (price, 2nd signature field).