✨(backend) admin api confirm quote of batch order
We need to give admin users the ability to confirm
that the organization has signed the quote.
When confirming the quote, we freeze the total and mark
the quote as signed. This step is crucial because,
regardless of the payment method, it allows the batch
order’s state to transition to the next step.
Since the product’s price is not applicable to batch
orders, organizations need to provide the price at which
they are going to sell the training to the company.
Without this total, the batch order’s quote cannot be confirmed.
This total is required because we create the main invoice
based on the price provided by the organization.