✨(backend) add contract definition to product for order and batch order
The contracts of a batch order are very different from
the contract of a single order. We can’t use the same
field to prepare both definitions in the Product model.
The appendices and the body are not the same at all.
We need admin users to be able to configure each
contract definition. One contract definition will be
for single orders (user-based), and the other contract
definition will be used when the offering is intended
for batch orders.