✨(backend) add contract definition to product for order and batch order
The contracts of the batch order are very different to the contract
of a single order. We can't use the same field to prepare both
definitions. The appendixes, and the body aren't the same at all.
We need our admin user to be able to configure each contract
definitions. The first contract definition will be for order (for user)
and the other contract definition should be set if the offering
aims batch orders.