✨(backend) professional training agreement for contract definition
A new template is needed to create the contract of a batch order.
There is now new two available choices for the template name
of a contract definition. Those two choices will be used of
the contract of a batch order. One is branded Unicamp, and
the other is without the brand. Factories and tests were updated
accordingly. For factories for single orders, contract definition
template will be used, and for batch order's contract the
professional training agreement template will be used.