✨(backend) batch order to get equivalent course run dates
Batch order was not yet ready to retrieve the course runs
of the target courses related to the product/course (offering).
This would  block us for the course section of the professional
agreement template. Now, we this method added, we can get more
information about the course start, end, and effort.