♻️(backend) prevent create batch order without contract definition
Before we would only control if the batch order is related to a contract
definition on the submitting to signature step. This would lead us to have
errors at a step where the batch order has already succeeded important
steps (quote generation, quote signature, freeze total and create
main invoice). To avoid having batch orders that are stuck in that
state, we decided to avoid the creation of a batch order if the product
does not have a contract definition for batch orders.