🔥(backend) remove submit for signature batch order client API
Since the batch order's owner will not be the one who
signs the contract, it was not anymore necessary to let
this endpoint exist. The signature procedure will be launched
from the organization's dashboard, and an email will be sent
to the batch order's signatory member with the invitation link
to sign the document.