♻️(backend) change batch order signatory for contract
The contract of a batch order will not be signed by the owner
of the object. This behavior is different from a single order.
In fact, we have a designated person from the buyer's side
of the batch order who is responsible to sign the contract.
We have updated the method in the SignatureBackend
LexPerson that prepares the buyer signer with this
person when it comes to a batch order.