💄(backend) render short dates in contracts templates
We want to render only the date without the hours into the
professional training agreement and the standard contract
templates. In the templates, we've changed the builtin template
tag to `SHORT_DATE_FORMAT` for the course start and end. Also,
to improve developer's experience, we have added the rendering
of the organization's logo when using the debug view of a
contract unicamp.