♻️(backend) use locale-aware short dates in contract context
Our contract templates would render the course dates into datetimes.
Although, the hours are not required to figure on the contract document.
We have decided to change the way we prepare the context of a contract
in the utils method. We now only use a short date format to prepare those
information. Those dates will be locale-aware depending on the language
code set in the contract definition of the product.