♻️(backend) prepare contract context per object type
When we prepare the contract's context of an order, we
don't necessary need the default values of the batch order
into the context that is saved into the object. We would
end up with a freezed context containing key and values
with default placeholders. Since the preparation of the
context method is used for orders and batch orders, we
have decided to only create the context needed for
each type of objects. Finally, our method that generates
the document context only prepares the awaited data
for each situation.