🧑‍💻(backend) debug view invitation link signature
While using the debug view of the invitation link signature
we would encounter many time errors related to the product
title already exists, or, the translation does not exist.
This was very painful. We have fixed this and we also
made the debug view locale aware with the parameter
`django-language` in your navigator cookies.