📌(backend) pin `SQLAlchemy` 2.0.40 and `sqlacodegen` 3.0.0
The dependency sqlacodegen depends on SQLAlchemy less than 2.0.42
version. There has been a bug report in the repository of sqlacodegen
3.1.x versions that requires a fix to support breaking change in
SQLAlchemy 2.0.42 version already. The upgrade of SQLAlchemy to 2.0.43
or 2.0.42 is not yet possible. We have to wait for a newer release of
sqlacodegen to be compatible with at least SQLAlchemy 2.0.42 version.
Issue: https://github.com/agronholm/sqlacodegen/issues/416
Sor for now we will pin those versions.