🐛(backend) orders with full discount vouchers
When an authenticated user applies a voucher with a full
discount rate, the init_flow method of the order would
not create a main invoice, even though the product has a fee.
This behavior caused issues when submitting the contract
for signature on orders for credential products.
We now ensure that if the order’s total is 0 because of a
full discount voucher, a main invoice is still created, since
the product is not free by default.