✨(backend) add `from_batch_order` to order serializer admin API
When an order was created from a batch order, the admin API
did not return any information indicating this, except for the
status `to_own` when it had not yet been assigned an owner.
We have decided to add a new key to the serializers
(`AdminOrderLightSerializer` and `AdminOrderSerializer`)
to explicitly indicate whether an order was created from a
batch order.