🐛(backend) freeze target courses for orders of batch order
We've forgotten to freeze the target courses of orders that were
generated from a batch order. Thus, it would create the order
but it would not enroll the student to the course run. We have
added this call in the method that generates the orders. Now,
when retrieving the list of orders, all the missing information
is present.