🐛(backend) set installment error in debit pending installment
When the payment provider has issues, we were supposed
to pass the installment id to the the method `set_installment_error`
instead of the installment object. This created issues in our sentry,
and we coudn't follow the fact that payment provider raised
an exception when calling `create_zero_click_payment`.