✨(backend) add payment method of batch order to quote serializer
For the API client, we have added the batch order's payment method in
the quote serializer. This value is important to understand in which
context the quote is in. If the batch order is paid by purchase order,
this information is required to show the button action for an
organization to confirm that it has received the purchase order
from the buyer. When it comes to the bank transfer payment method,
it will allow the organization to confirm once it has received
the payment.