♻️(backend) contract information for batch order serializer admin API
We previously only return the `contract_id` in the response of
the batch order admin serializer. The detail view of the batch
order in the backoffice should render more information about
the contract.

With this update, it returns the same amount of information as
the admin Order serializer that is used for a single order.