♻️(backend) admin order serializer batch order field
We updated the order serializer response to include more
meaningful information about whether an order was
generated from a batch order.

Previously, we only returned a boolean field
from_batch_order, but it is now more appropriate to return
the id of the batch order from which the order was created.