♻️(backend) refactor invoice reference
There was a bug when refactoring the `BatchOrderFactory`
for the creation of the invoice when the state is `completed`.
The Invoice Factory would create an order that was not meant
to be created. This would lead some tests concerning orders list
to add an extra order when we prepare orders from batch orders.

The issue is that the timezone timestamp would return the same
value while generating the reference of an invoice.