♻️(backend) batch order get signature invitation link
When using the `DummyBackendSignature` with batch orders,
we have added an automatic buyer signature when the
API consumer requests the invitation link while submitting
the contract to the signature provider. When the organization
signs the contract, the standard behavior is kept, letting
the API consumer to request the notification event endpoint.

This allows locally for developers to simply request to
submit for signature the contract for the buyer instead
of making a separate POST request to the signature
notification handler endpoint. We did this because
the buyer signs the contract from an email invitation.

Also, we have remove a call of batch order flow update
within the `submit_for_signature` method because the
state of the batch order was already in `to_sign`
when calling it. It was useless.