♻️(backend) update `contracts_signature_link` for batch orders
The method `get_abilities` on the Contract model was not yet updated
for user's who have accesses to organization concerning bathc orders
We've now updated the permissions of this method to check whether
or not the user has the authorization to sign the contracts
related to batch order of his organization.

Also, we have updated the method 'contracts_signature_link' to
be able to retrieve only batch order contracts. Organization users
with owner role can now filter by contract ids or offering ids to
retrieve a signature link to sign in bulk contracts.

Fix #1237