✨(backend) add agreement batch order viewset and serializer
We want authenticated organization user who have access to be able to
list, retrieve information on the contracts related to batch orders. This
will allow certain user who have the rights to retrieve the list of contracts
and if they are eligible to sign them.

We needed to create a new viewset and serializer because the existing
viewset for organization nested contract viewset was built around the
relation `contract.order`, it did not correspond to the relation
`batch_order.contract`. This is the reason why we created a new viewset
for that matter.