✨(backend) add process payment state for batch order
We want to have a distinct state when the batch order is ready for
payment and when it has been submitted to payment. This is the
reason we've introduced into the flow state of batch order
`process_payment`.

This state can only be reached by batch orders that are using
the `card_payment` payment method. While implementing this
new state, now the payment method `bank_transfer` and
`card_payment` transitions to `pending` once the contract is
signed by the buyer. We have changed this to allow the API
consumers to know which step is the next according to the
cycle of a batch order.