✨(backend) add `organization_signed_on` in agreements serializer
The frontend client API consumers needed some information
about the signature state of an organization on a contract for batch
order. From their perspective, they need to know if the organization
has signed the contract yet.

For testing purposes, we also added in the BatchOrderFactory
that the batch order is signed by the organization when we set
the state to `completed`.