✨(backend) invitation link to signatory email of batch order
We made an update to send the invitation link to sign the agreement
to the signatory representative of the batch order. We changed it
because the batch order owner is not the person who will sign the
agreement all the time.

The preparation of the signature procedure is prepare for that
specific person already ( signatory representative ) when the
contract is related to a batch order.