✨(backend) signatory batch order fields required
We want our signatory fields of a batch required because they
have a crucial role into signing the contract of the order. Before,
it was only controlled at the API level that those fields were
required, but not to enhance the control, we also made them
required at the model level.