♻️(backend) organization agreement viewset exclude `failed_payment`
For the agreement organization viewset, we don't want authenticated user
with all access to the organization to see contracts that are canceled or in
failed payment. This view is used by our API consumers to sign in bulk the
contracts. This is the reason we don't want organization to sign contracts
that are canceled or with payment issues.