🎨(backend) conditionally render elements based quote context
Into the quote template, when the funding entity was not
given on the batch order, it would render `None` instead
of a a '-'. The same as been applied for the funding amount
where a 0 would be rendered instead of '-'. We also added
the currency symbol if a funding amount was declared into
the batch order creation. We've also fix some extra
whitespace that were not needed.