🐛(backend) admin backoffice batch order completed gets canceled
In the frontend backoffice of Joanie, once a batch order with
generated orders was canceled, the property vouchers would
return an error and thus, the detail view of a batch order will
not render any result. The issue came that it was searching
for the attribute `code` of a NoneType object because the
voucher related to the order `to_own` was deleted.