♻️(backend) batch order failed_payment state only for card payment
When looking over the code again, we have considered that only
failed payment can happen when the payment method is with
card_payment. Because it's the payment backend that notifies
us that the payment has failed. Whereas, with the two other
payment methods, the payment is done outside the application.
So until the payment is not verified outside, the organization
should not trigger the next action to consider the
batch order is paid.